Bound the SM120 CTA tile and widen the sparse batch stride
The SM120 dense builder bounds no mode of the CTA tile against the MMA tiler.
The mainloop steps the tiled MMA with a ceiling division, thus a K mode that is
not a multiple of 32 makes the mainloop read past the shared memory tile.
compute-sanitizer reports an invalid shared read of 16 bytes for a tile of
128x64x48 over a problem that holds more than one K tile.

The blockwise builder is the only one of six SM120 mainloop builders with no
cluster bound. The new line copies the five siblings.

The sparse batch stride is a product of two int extents, and the cast to int64_t
widens a value that already wrapped. Eight sites, four in each configuration
header. At M and K of 65536 the stride reaches exactly 0, thus each batch reads
one address.

+++ include/cutlass/gemm/collective/builders/sm1xx_sparse_config.inl
@@ -266,7 +266,7 @@ struct Sm1xxGemmSparseConfig {
                    int32_t(L)),
         make_stride(int64_t(K_AlignedAC),
                     make_stride(_1{}, ElementASparsity{}),
-                    (L == 1) ? int64_t(0) : int64_t(M_AlignedAC * K_AlignedAC))
+                    (L == 1) ? int64_t(0) : int64_t(M_AlignedAC) * int64_t(K_AlignedAC))
       );
     }
     else {
@@ -276,7 +276,7 @@ struct Sm1xxGemmSparseConfig {
                    int32_t(L)),
         make_stride(ElementASparsity{},
                     make_stride(_1{}, int64_t(M_AlignedAC) * ElementASparsity{}),
-                    (L == 1) ? int64_t(0) : int64_t(M_AlignedAC * K_AlignedAC))
+                    (L == 1) ? int64_t(0) : int64_t(M_AlignedAC) * int64_t(K_AlignedAC))
       );
     }
   }
@@ -302,8 +302,8 @@ struct Sm1xxGemmSparseConfig {
                  make_shape(shape<1>(TensorEAtom{}), int32_t(K_AlignedE / TensorEAtomK{})),
                  int32_t(L)),
       make_stride(make_stride(stride<0>(TensorEAtom{}), cute::Int<cute::cosize(TensorEAtom{})>{}),
-                  make_stride(stride<1>(TensorEAtom{}), int64_t(M_AlignedE * TensorEAtomK{})),
-                  (L == 1) ? int64_t(0) : int64_t(M_AlignedE * K_AlignedE))
+                  make_stride(stride<1>(TensorEAtom{}), int64_t(M_AlignedE) * int64_t(TensorEAtomK{})),
+                  (L == 1) ? int64_t(0) : int64_t(M_AlignedE) * int64_t(K_AlignedE))
     );
   }
 };
